Extract TextWriter & RTFWriter from StyledText (issue #292)
